EDGE-1397: Add 'decryptable' endpoint to sovereign
This allows testing if a provided encrypted string is decryptable by
the server's configured key.

The need for this arises out of a deployment with multiple environments,
and no clear way to distinguish an encrypted blob intended for one
environment with another.

Eg. A secret is encrypted with the dev sovereign instance, and a separate
one is encrypted with the staging sovereign instance. The blobs of text
have no indicator that shows which environment they are meant to be
decrypted in.

Using this endpoint, you can pass the encrypted blob to check if it will
work on the sovereign instance. This can be added to a build or deployment
pipeline, to ensure that the deployment doesn't result in a decryption failure.