Still attempting to fix log issues
Some logs seem to have partial fields which don't make sense.
Such as, the method field is missing, but the resources requested
is not missing.
This indicates to me some kind of concurrency problem.
I'm not really sure how to deal with it, but hopefully by adding
the formatting/merging to the list of structlog processors, in the
correct order... it should result in the correct dictionary format.