fix: register with the worker, async, try many times
registering directly was good, and supposedly takes a load off the
workers shoulders. but registering directly has the downside of not
triggering a render job for that request at the same time.

this means that the web can get a cache entry from s3, write it back to
the local cache, and then that local entry can persist for as long as
context does not trigger a re-render of that client. so although the
client is registered, the local entry will be served until such time
that a render is triggered. For a template which depends on very few
context, which change very infrequently, this could mean that the cache
entry can become stale and invalid while everything else around it
changes.